Turning this into library (Phase 1) - Removed all TwitchSpy and Mongo stuff (will be in separate repo) - Google Tests added - Lot's of tests - Lot's of improvements

+++ TwitchXX/TwitchChannelFeed.h
@@ -11,11 +11,11 @@ namespace TwitchXX
 		virtual ~TwitchChannelFeed();
 		
 		TwitchPostsContainer GetPosts(const std::wstring& channel_name, size_t count = 0) const;
-		TwitchPost GetPost(const std::wstring& channel_name, size_t id) const; //TODO: size_t is not quite right. May be will had to change to string type
+		TwitchPost GetPost(const std::wstring& channel_name, unsigned long long id) const; //TODO: size_t is not quite right. May be will had to change to string type
 		TwitchPost Post(const std::wstring& channel_name, const std::wstring& body, bool share = false) const;
-		bool DeletePost(const std::wstring& channel_name, size_t id) const; //TODO: size_t -> string?
-		bool AddReaction(const std::wstring& channel_name, size_t id, size_t emote_id) const;
-		bool RemoveReaction(const std::wstring& channel_name, size_t id, size_t emote_id) const;
+		bool DeletePost(const std::wstring& channel_name, unsigned long long id) const; //TODO: size_t -> string?
+		bool AddReaction(const std::wstring& channel_name, unsigned long long id, size_t emote_id) const;
+		bool RemoveReaction(const std::wstring& channel_name, unsigned long long id, size_t emote_id) const;
